Refactored HTML value processing
To allow grid values to have multiple HTML values or on a different property (e.g. combined image and RTE), the processing is moved from the control to the value. If the editor configuration is needed for processing the HTML, it should be injected before returning the value (see TextstringGridControl).